+++ src/wrapper/executor.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use crate::{errors::*, objects::JObject, JNIEnv, JavaVM};
+use crate::{errors::*, JNIEnv, JavaVM};
 
 /// The capacity of local frames, allocated for attached threads by default. Same as the default
 /// value Hotspot uses when calling native Java methods.
@@ -39,9 +39,7 @@ pub const DEFAULT_LOCAL_FRAME_CAPACITY: i32 = 32;
 ///
 /// let val: jint = exec.with_attached(|env| {
 ///    let x = JValue::from(-10);
-///    let val: jint = env.call_static_method("java/lang/Math", "abs", "(I)I", &[x])?
-///      .i()?;
-///    Ok(val)
+///    env.call_static_method("java/lang/Math", "abs", "(I)I", &[x])?.i()
 /// })?;
 ///
 /// assert_eq!(val, 10);
@@ -68,20 +66,15 @@ impl Executor {
     /// call.
     ///
     /// Allocates a local frame with the specified capacity.
-    pub fn with_attached_capacity<F, R>(&self, capacity: i32, f: F) -> Result<R>
+    pub fn with_attached_capacity<F, T, E>(&self, capacity: i32, f: F) -> std::result::Result<T, E>
     where
-        F: FnOnce(&mut JNIEnv) -> Result<R>,
+        F: FnOnce(&mut JNIEnv) -> std::result::Result<T, E>,
+        E: From<Error>,
     {
         assert!(capacity > 0, "capacity should be a positive integer");
 
         let mut jni_env = self.vm.attach_current_thread_as_daemon()?;
-        let mut result = None;
-        jni_env.with_local_frame(capacity, |jni_env| {
-            result = Some(f(jni_env));
-            Ok(JObject::null())
-        })?;
-
-        result.expect("The result should be Some or this line shouldn't be reached")
+        jni_env.with_local_frame(capacity, |jni_env| f(jni_env))
     }
 
     /// Executes a provided closure, making sure that the current thread
@@ -90,9 +83,10 @@ impl Executor {
     ///
     /// Allocates a local frame with
     /// [the default capacity](constant.DEFAULT_LOCAL_FRAME_CAPACITY.html).
-    pub fn with_attached<F, R>(&self, f: F) -> Result<R>
+    pub fn with_attached<F, T, E>(&self, f: F) -> std::result::Result<T, E>
     where
-        F: FnOnce(&mut JNIEnv) -> Result<R>,
+        F: FnOnce(&mut JNIEnv) -> std::result::Result<T, E>,
+        E: From<Error>,
     {
         self.with_attached_capacity(DEFAULT_LOCAL_FRAME_CAPACITY, f)
     }